backend/requirements.txt
fastapi==0.133.0
uvicorn[standard]==0.41.0
sqlalchemy==2.0.47
alembic==1.18.4
psycopg2-binary==2.9.11
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.22
httpx==0.28.1
